moved all n queens concept implementations into one file-all_n_queens
